requirements.txt
asgiref==3.4.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
Django==3.2.9
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-mysql==4.5.0
idna==2.10
numpy==1.22.3
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyMySQL==1.0.2
pytz==2020.4
qrcode==7.3.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
